+++ CHANGELOG.md
@@ -7,3 +7,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 - Initial release
+- Added Lexicon view to manage lexicon entries from the first namespace in the model